Renamed DismissibleNotification to LeftPaneNotification
